requirements.txt
flask==1.1.2
python-dotenv==0.12.0
sqlalchemy==1.3.16
flask-sqlalchemy==2.4.1
sqlalchemy-utils==0.36.3
flask-wtf==0.14.3
requests==2.23.0
redis==3.4.1
celery==4.4.2
gunicorn==20.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pynsq==0.8.3
dill==0.3.1.1